requirements.txt
aiosqlite==0.19.0
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
apprise==1.6.0
asgi-lifespan==2.1.0
asttokens==2.4.1
asyncpg==0.28.0
attrs==23.1.0
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cleo==2.1.0
click==8.1.7
cloudpickle==3.0.0
clr-loader==0.2.6
colorama==0.4.6
comm==0.1.4
coolname==2.2.0
crashtest==0.4.1
croniter==1.4.1
cryptography==41.0.5
cssselect==1.2.0
cssutils==2.9.0
dateparser==1.1.8
db-dtypes==1.1.1
dbharbor==0.1.4
debugpy==1.8.0
decorator==5.1.1
demail @ git+https://github.com/edire/demail.git@7e614361f729f5f44b7e461e2cb77f1ed0806faa
distlib==0.3.8
dlogging @ git+https://github.com/edire/dlogging.git@12c73a61a71ebde99c8e078df3a663ca78fb116a
docker==6.1.3
dulwich==0.21.7
dwebdriver @ git+https://github.com/edire/dwebdriver.git@517c8ac14250ddb7e2d2d8791e9c1a420f77f595
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fsspec==2023.10.0
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
graphviz==0.20.1
greenlet==3.0.0
griffe==0.36.7
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==7.0.1
installer==0.7.0
ipykernel==6.26.0
ipython==8.17.2
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.5.0
jupyter_core==5.5.0
keyring==24.3.0
kubernetes==27.2.0
lxml==4.9.3
Mako==1.2.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
msal==1.24.1
msgpack==1.0.7
nest-asyncio==1.5.8
numpy==1.26.1
oauthlib==3.2.2
Office365-REST-Python-Client==2.5.1
openpyxl==3.1.2
orjson==3.9.9
outcome==1.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pathspec==0.11.2
pendulum==2.1.2
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prefect==2.13.8
premailer==3.10.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyodbc==5.0.1
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-slugify==8.0.1
pythonnet==3.0.3
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.6.1
readchar==4.0.5
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
ruamel.yaml==0.18.1
ruamel.yaml.clib==0.2.8
sageintacctsdk @ git+https://github.com/edire/sageintacct.git@e10c9665fe3d624cd6f2b5b0d620fb014d50e35c
SecretStorage==3.3.3
selenium==4.14.0
sendgrid==6.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.22
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.27.0
text-unidecode==1.3
toml==0.10.2
tomlkit==0.12.3
tornado==6.3.3
traitlets==5.13.0
trio==0.22.2
trio-websocket==0.11.1
trove-classifiers==2023.11.29
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.23.2
virtualenv==20.25.0
wcwidth==0.2.9
webdriver-manager==4.0.1
websocket-client==1.6.4
websockets==12.0
wsproto==1.2.0
xmltodict==0.12.0
yagmail==0.15.293
zipp==3.17.0
